Refactor Oracle loader utilities
