+++ src/main/java/com/example/h2sync/service/AbstractOracleLoaderService.java
@@ -8,14 +8,7 @@ import org.springframework.jdbc.datasource.DriverManagerDataSource;
 import javax.sql.DataSource;
 import java.io.IOException;
 import java.io.Reader;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.sql.*;
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.OffsetDateTime;
-import java.time.ZonedDateTime;
 import java.util.*;
 import java.util.concurrent.*;
 import java.util.regex.Matcher;
@@ -32,6 +25,7 @@ public abstract class AbstractOracleLoaderService {
     protected final int maxRetries;
     protected final Set<String> blacklist;
     protected final String oracleSchema;
+    private final OracleViewSqlTranslator viewSqlTranslator;
 
     protected AbstractOracleLoaderService(
             JdbcTemplate h2,
@@ -48,6 +42,7 @@ public abstract class AbstractOracleLoaderService {
         this.batchSize = batchSize;
         this.maxRetries = maxRetries;
         this.oracleSchema = schema != null ? schema.toUpperCase(Locale.ROOT) : null;
+        this.viewSqlTranslator = new OracleViewSqlTranslator(this.oracleSchema);
 
         this.blacklist = Arrays.stream((blacklistCsv == null ? "" : blacklistCsv).split(","))
                 .map(String::trim)
@@ -357,7 +352,7 @@ public abstract class AbstractOracleLoaderService {
         log.info("Creating H2 view {} from Oracle view {}", viewName, src);
 
         String oracleSql = fetchOracleViewDefinition(view);
-        String translatedSql = translateViewSql(oracleSql);
+        String translatedSql = viewSqlTranslator.translate(oracleSql);
 
         dropLegacyArtifacts(viewName, view);
 
@@ -559,361 +554,6 @@ public abstract class AbstractOracleLoaderService {
         }
     }
 
-    private String translateViewSql(String oracleSql) {
-        if (oracleSql == null) {
-            throw new IllegalArgumentException("Oracle view SQL is null");
-        }
-        String stripped = stripSqlTerminator(stripComments(oracleSql)).trim();
-        if (stripped.isEmpty()) {
-            throw new IllegalArgumentException("Oracle view SQL is empty after cleaning");
-        }
-        String withoutSchema = removeOracleSchemaQualifiers(stripped);
-        String cleaned = removeTrailingReadOnly(withoutSchema).trim();
-        String uppercased = uppercaseUnquotedIdentifiers(cleaned);
-        String protectedKeywords = quoteH2ReservedKeywords(uppercased);
-        return replaceOracleSpecificFunctions(protectedKeywords);
-    }
-
-    private String stripSqlTerminator(String sql) {
-        String out = sql.trim();
-        while (out.endsWith(";") || out.endsWith("/")) {
-            out = out.substring(0, out.length() - 1).trim();
-        }
-        return out;
-    }
-
-    private String stripComments(String sql) {
-        String noBlock = sql.replaceAll("(?is)/\\*.*?\\*/", " ");
-        StringBuilder sb = new StringBuilder();
-        try (Scanner scanner = new Scanner(noBlock)) {
-            while (scanner.hasNextLine()) {
-                String line = scanner.nextLine();
-                int idx = line.indexOf("--");
-                if (idx >= 0) {
-                    line = line.substring(0, idx);
-                }
-                if (!line.isBlank()) {
-                    if (sb.length() > 0) {
-                        sb.append('\n');
-                    }
-                    sb.append(line);
-                }
-            }
-        }
-        return sb.toString();
-    }
-
-    private String removeOracleSchemaQualifiers(String sql) {
-        if (oracleSchema == null || oracleSchema.isBlank()) {
-            return sql;
-        }
-        String result = sql;
-        String quotedSchema = Pattern.quote("\"" + oracleSchema + "\"");
-        result = result.replaceAll("(?i)" + quotedSchema + "\\.", "");
-        result = result.replaceAll("(?i)" + Pattern.quote(oracleSchema) + "\\.", "");
-        return result;
-    }
-
-    private String removeTrailingReadOnly(String sql) {
-        return sql.replaceAll("(?i)WITH\\s+READ\\s+ONLY", " ")
-                .replaceAll("(?i)WITH\\s+CHECK\\s+OPTION", " ");
-    }
-
-    private String uppercaseUnquotedIdentifiers(String sql) {
-        StringBuilder result = new StringBuilder(sql.length());
-        int len = sql.length();
-        int i = 0;
-        while (i < len) {
-            char c = sql.charAt(i);
-            if (c == '\'') {
-                result.append(c);
-                i++;
-                while (i < len) {
-                    char current = sql.charAt(i);
-                    result.append(current);
-                    if (current == '\'') {
-                        if (i + 1 < len && sql.charAt(i + 1) == '\'') {
-                            result.append(sql.charAt(i + 1));
-                            i += 2;
-                            continue;
-                        }
-                        i++;
-                        break;
-                    }
-                    i++;
-                }
-            } else if (c == '"') {
-                result.append(c);
-                i++;
-                while (i < len) {
-                    char current = sql.charAt(i);
-                    result.append(current);
-                    if (current == '"') {
-                        if (i + 1 < len && sql.charAt(i + 1) == '"') {
-                            result.append(sql.charAt(i + 1));
-                            i += 2;
-                            continue;
-                        }
-                        i++;
-                        break;
-                    }
-                    i++;
-                }
-            } else if (Character.isLetter(c) || c == '_' || c == '$' || c == '#') {
-                int start = i;
-                i++;
-                while (i < len) {
-                    char current = sql.charAt(i);
-                    if (Character.isLetterOrDigit(current) || current == '_' || current == '$' || current == '#') {
-                        i++;
-                    } else {
-                        break;
-                    }
-                }
-                result.append(sql.substring(start, i).toUpperCase(Locale.ROOT));
-            } else {
-                result.append(c);
-                i++;
-            }
-        }
-        return result.toString();
-    }
-
-    private static final Set<String> H2_RESERVED_IDENTIFIERS =
-            Set.of("VALUE", "TYPE");
-
-    private String quoteH2ReservedKeywords(String sql) {
-        StringBuilder result = new StringBuilder(sql.length() + 16);
-        int len = sql.length();
-        int i = 0;
-        while (i < len) {
-            char c = sql.charAt(i);
-            if (c == '\'') {
-                result.append(c);
-                i++;
-                while (i < len) {
-                    char current = sql.charAt(i);
-                    result.append(current);
-                    if (current == '\'') {
-                        if (i + 1 < len && sql.charAt(i + 1) == '\'') {
-                            result.append(sql.charAt(i + 1));
-                            i += 2;
-                            continue;
-                        }
-                        i++;
-                        break;
-                    }
-                    i++;
-                }
-            } else if (c == '"') {
-                result.append(c);
-                i++;
-                while (i < len) {
-                    char current = sql.charAt(i);
-                    result.append(current);
-                    if (current == '"') {
-                        if (i + 1 < len && sql.charAt(i + 1) == '"') {
-                            result.append(sql.charAt(i + 1));
-                            i += 2;
-                            continue;
-                        }
-                        i++;
-                        break;
-                    }
-                    i++;
-                }
-            } else if (Character.isLetter(c) || c == '_' || c == '$' || c == '#') {
-                int start = i;
-                i++;
-                while (i < len) {
-                    char current = sql.charAt(i);
-                    if (Character.isLetterOrDigit(current) || current == '_' || current == '$' || current == '#') {
-                        i++;
-                    } else {
-                        break;
-                    }
-                }
-                String token = sql.substring(start, i);
-                if (H2_RESERVED_IDENTIFIERS.contains(token)) {
-                    result.append('"').append(token).append('"');
-                } else {
-                    result.append(token);
-                }
-            } else {
-                result.append(c);
-                i++;
-            }
-        }
-        return result.toString();
-    }
-
-    private String replaceOracleSpecificFunctions(String sql) {
-        return replaceNvl2(sql);
-    }
-
-    private String replaceNvl2(String sql) {
-        StringBuilder result = new StringBuilder(sql.length());
-        int len = sql.length();
-        int i = 0;
-        while (i < len) {
-            if (matchesFunction(sql, i, "NVL2")) {
-                int funcStart = i;
-                i += 4;
-                int afterName = skipWhitespace(sql, i);
-                if (afterName >= len || sql.charAt(afterName) != '(') {
-                    result.append(sql, funcStart, i);
-                    continue;
-                }
-                int openParenIndex = afterName;
-                int closeParenIndex = findMatchingParenthesis(sql, openParenIndex);
-                if (closeParenIndex == -1) {
-                    result.append(sql, funcStart, i);
-                    continue;
-                }
-                String inner = sql.substring(openParenIndex + 1, closeParenIndex);
-                List<String> args = splitTopLevelArguments(inner);
-                if (args.size() != 3) {
-                    result.append(sql, funcStart, closeParenIndex + 1);
-                    i = closeParenIndex + 1;
-                    continue;
-                }
-                String expr = replaceNvl2(args.get(0).trim());
-                String notNullValue = replaceNvl2(args.get(1).trim());
-                String nullValue = replaceNvl2(args.get(2).trim());
-                result.append("CASE WHEN ")
-                        .append(expr)
-                        .append(" IS NOT NULL THEN ")
-                        .append(notNullValue)
-                        .append(" ELSE ")
-                        .append(nullValue)
-                        .append(" END");
-                i = closeParenIndex + 1;
-            } else {
-                result.append(sql.charAt(i));
-                i++;
-            }
-        }
-        return result.toString();
-    }
-
-    private boolean matchesFunction(String sql, int offset, String name) {
-        int len = sql.length();
-        int nameLen = name.length();
-        if (offset + nameLen > len) {
-            return false;
-        }
-        for (int i = 0; i < nameLen; i++) {
-            char c = sql.charAt(offset + i);
-            if (Character.toUpperCase(c) != name.charAt(i)) {
-                return false;
-            }
-        }
-        if (offset > 0) {
-            char prev = sql.charAt(offset - 1);
-            if (Character.isLetterOrDigit(prev) || prev == '_' || prev == '$' || prev == '#') {
-                return false;
-            }
-        }
-        if (offset + nameLen < len) {
-            char next = sql.charAt(offset + nameLen);
-            if (Character.isLetterOrDigit(next) || next == '_' || next == '$' || next == '#') {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    private int skipWhitespace(String sql, int index) {
-        int len = sql.length();
-        int i = index;
-        while (i < len && Character.isWhitespace(sql.charAt(i))) {
-            i++;
-        }
-        return i;
-    }
-
-    private int findMatchingParenthesis(String sql, int openIndex) {
-        int len = sql.length();
-        int depth = 0;
-        boolean inSingle = false;
-        boolean inDouble = false;
-        for (int i = openIndex; i < len; i++) {
-            char c = sql.charAt(i);
-            if (c == '\'' && !inDouble) {
-                if (inSingle && i + 1 < len && sql.charAt(i + 1) == '\'') {
-                    i++;
-                } else {
-                    inSingle = !inSingle;
-                }
-                continue;
-            }
-            if (c == '"' && !inSingle) {
-                if (inDouble && i + 1 < len && sql.charAt(i + 1) == '"') {
-                    i++;
-                } else {
-                    inDouble = !inDouble;
-                }
-                continue;
-            }
-            if (inSingle || inDouble) {
-                continue;
-            }
-            if (c == '(') {
-                depth++;
-            } else if (c == ')') {
-                depth--;
-                if (depth == 0) {
-                    return i;
-                }
-                if (depth < 0) {
-                    return -1;
-                }
-            }
-        }
-        return -1;
-    }
-
-    private List<String> splitTopLevelArguments(String input) {
-        List<String> parts = new ArrayList<>();
-        int len = input.length();
-        int start = 0;
-        int depth = 0;
-        boolean inSingle = false;
-        boolean inDouble = false;
-        for (int i = 0; i < len; i++) {
-            char c = input.charAt(i);
-            if (c == '\'' && !inDouble) {
-                if (inSingle && i + 1 < len && input.charAt(i + 1) == '\'') {
-                    i++;
-                } else {
-                    inSingle = !inSingle;
-                }
-                continue;
-            }
-            if (c == '"' && !inSingle) {
-                if (inDouble && i + 1 < len && input.charAt(i + 1) == '"') {
-                    i++;
-                } else {
-                    inDouble = !inDouble;
-                }
-                continue;
-            }
-            if (inSingle || inDouble) {
-                continue;
-            }
-            if (c == '(') {
-                depth++;
-            } else if (c == ')') {
-                depth--;
-            } else if (c == ',' && depth == 0) {
-                parts.add(input.substring(start, i));
-                start = i + 1;
-            }
-        }
-        parts.add(input.substring(start));
-        return parts;
-    }
-
     private void createTargetTableFrom(ResultSetMetaData md, String target) throws SQLException {
         String drop = "DROP TABLE IF EXISTS " + target + " CASCADE";
         h2.execute(drop);
@@ -1024,7 +664,7 @@ public abstract class AbstractOracleLoaderService {
                             long n = 0;
                             while (rs.next()) {
                                 for (int i = 1; i <= cols; i++) {
-                                    Object v = readColumnValue(rs, md, i);
+                                    Object v = OracleJdbcValueConverter.readColumnValue(rs, md, i, log);
                                     ins.setObject(i, v);
                                 }
                                 ins.addBatch();
@@ -1056,204 +696,5 @@ public abstract class AbstractOracleLoaderService {
         }
     }
 
-    private Object readColumnValue(ResultSet rs, ResultSetMetaData md, int index) throws SQLException {
-        int jdbcType = md.getColumnType(index);
 
-        if (isBinaryType(jdbcType)) {
-            byte[] data = rs.getBytes(index);
-            return rs.wasNull() ? null : data;
-        }
-
-        if (isClobType(jdbcType)) {
-            String text = rs.getString(index);
-            return rs.wasNull() ? null : text;
-        }
-
-        if (isTemporalType(jdbcType)) {
-            try {
-                Timestamp ts = rs.getTimestamp(index);
-                if (ts != null || rs.wasNull()) {
-                    return ts;
-                }
-            } catch (SQLException ex) {
-                log.debug("Oracle timestamp conversion failed for column {} using getTimestamp: {}", index, ex.getMessage());
-            }
-        }
-
-        Object value = rs.getObject(index);
-
-        if (value instanceof Blob) {
-            return blobToBytes((Blob) value);
-        }
-        if (value instanceof Clob) {
-            return clobToString((Clob) value);
-        }
-
-        if (isTemporalType(jdbcType)) {
-            return toTimestamp(value);
-        }
-        return maybeConvertTemporal(value);
-    }
-
-    private boolean isTemporalType(int jdbcType) {
-        return jdbcType == Types.DATE
-                || jdbcType == Types.TIME
-                || jdbcType == Types.TIMESTAMP
-                || jdbcType == Types.TIMESTAMP_WITH_TIMEZONE
-                || jdbcType == Types.TIME_WITH_TIMEZONE;
-    }
-
-    private boolean isBinaryType(int jdbcType) {
-        return jdbcType == Types.BLOB
-                || jdbcType == Types.BINARY
-                || jdbcType == Types.VARBINARY
-                || jdbcType == Types.LONGVARBINARY;
-    }
-
-    private boolean isClobType(int jdbcType) {
-        return jdbcType == Types.CLOB
-                || jdbcType == Types.NCLOB
-                || jdbcType == Types.LONGVARCHAR
-                || jdbcType == Types.LONGNVARCHAR;
-    }
-
-    private Object maybeConvertTemporal(Object value) throws SQLException {
-        if (value == null) return null;
-        if (value instanceof Timestamp) return value;
-        if (value instanceof java.sql.Date || value instanceof java.sql.Time || value instanceof java.util.Date) {
-            return toTimestamp(value);
-        }
-        if (value instanceof LocalDateTime || value instanceof LocalDate
-                || value instanceof Instant || value instanceof OffsetDateTime
-                || value instanceof ZonedDateTime) {
-            return toTimestamp(value);
-        }
-        String className = value.getClass().getName();
-        if (className.startsWith("oracle.sql.")) {
-            return toTimestamp(value);
-        }
-        return value;
-    }
-
-    private byte[] blobToBytes(Blob blob) throws SQLException {
-        if (blob == null) {
-            return null;
-        }
-        try {
-            long length = blob.length();
-            if (length > Integer.MAX_VALUE) {
-                throw new SQLException("BLOB length exceeds supported size: " + length);
-            }
-            return blob.getBytes(1, (int) length);
-        } finally {
-            free(blob);
-        }
-    }
-
-    private String clobToString(Clob clob) throws SQLException {
-        if (clob == null) {
-            return null;
-        }
-        try {
-            long length = clob.length();
-            if (length > Integer.MAX_VALUE) {
-                throw new SQLException("CLOB length exceeds supported size: " + length);
-            }
-            return clob.getSubString(1, (int) length);
-        } finally {
-            free(clob);
-        }
-    }
-
-    private void free(Blob blob) {
-        if (blob == null) {
-            return;
-        }
-        try {
-            blob.free();
-        } catch (AbstractMethodError | SQLException ignored) {
-            // Some drivers (or older JDBC versions) do not support free(); ignore such cases.
-        }
-    }
-
-    private void free(Clob clob) {
-        if (clob == null) {
-            return;
-        }
-        try {
-            clob.free();
-        } catch (AbstractMethodError | SQLException ignored) {
-            // Ignore drivers that do not implement free()
-        }
-    }
-
-    private Object toTimestamp(Object value) throws SQLException {
-        if (value == null) return null;
-        if (value instanceof Timestamp) return value;
-        if (value instanceof java.sql.Date) {
-            return new Timestamp(((java.sql.Date) value).getTime());
-        }
-        if (value instanceof java.sql.Time) {
-            return new Timestamp(((java.sql.Time) value).getTime());
-        }
-        if (value instanceof java.util.Date) {
-            return new Timestamp(((java.util.Date) value).getTime());
-        }
-        if (value instanceof LocalDateTime) {
-            return Timestamp.valueOf((LocalDateTime) value);
-        }
-        if (value instanceof LocalDate) {
-            return Timestamp.valueOf(((LocalDate) value).atStartOfDay());
-        }
-        if (value instanceof Instant) {
-            return Timestamp.from((Instant) value);
-        }
-        if (value instanceof OffsetDateTime) {
-            return Timestamp.valueOf(((OffsetDateTime) value).toLocalDateTime());
-        }
-        if (value instanceof ZonedDateTime) {
-            return Timestamp.valueOf(((ZonedDateTime) value).toLocalDateTime());
-        }
-
-        String className = value.getClass().getName();
-        if (className.startsWith("oracle.sql.")) {
-            try {
-                Method method = value.getClass().getMethod("timestampValue");
-                Object ts = method.invoke(value);
-                if (ts instanceof Timestamp) {
-                    return ts;
-                }
-            } catch (NoSuchMethodException ignored) {
-                // fall through and try other conversion methods
-            } catch (IllegalAccessException | InvocationTargetException e) {
-                throw new SQLException("Failed to convert Oracle temporal value of type " + className, e);
-            }
-
-            try {
-                Method method = value.getClass().getMethod("dateValue");
-                Object date = method.invoke(value);
-                if (date instanceof java.sql.Date) {
-                    return new Timestamp(((java.sql.Date) date).getTime());
-                }
-            } catch (NoSuchMethodException ignored) {
-                // fall through
-            } catch (IllegalAccessException | InvocationTargetException e) {
-                throw new SQLException("Failed to convert Oracle temporal value of type " + className, e);
-            }
-
-            try {
-                Method method = value.getClass().getMethod("timeValue");
-                Object time = method.invoke(value);
-                if (time instanceof java.sql.Time) {
-                    return new Timestamp(((java.sql.Time) time).getTime());
-                }
-            } catch (NoSuchMethodException ignored) {
-                // fall through
-            } catch (IllegalAccessException | InvocationTargetException e) {
-                throw new SQLException("Failed to convert Oracle temporal value of type " + className, e);
-            }
-        }
-
-        return value;
-    }
 }
